Implement PicassoModule and ApplicationModule
If any need for application or its context(For this case Picasso needs
a context to be provided), it needs to be provided by a module.
The module should be accept the application as a constructor
parameter. So, when building the dagger component in the application,
the module should be created manually by passing application, then
the function that is generated by dagger should be called with the module.