Inject a dependency from ApplicationComponent into an Activity
A function that accepts an android related component as a parameter
(in this case it is Activity) and name is inject by convention should be
created in the ApplicationComponent.

In the Activity this function should be used. To access the function,
the activity should access the application and get the component instance.
Therefore, The function can be called on the component by passing 'this'.

As a conclusion, any dependency from ApplicationComponent can be injected
into activity by using @Inject keyword(in this case it is Picasso)