Share a single instance between Activity and Fragments
In this case the shared object is HomeViewModel's. Since it is marked
with @Singleton, dagger can serve it as a single instance.

In this design inject functions for Activity and Fragments are written
in ApplicationComponent. For fragments inject functions should be called
in onAttach function. inject function could be called either before super
or after.

While we are defining the ViewModel as Singleton, after logout, the temporary
state still stays on the memory. So, login with another user is encountered
with the last users state. This is not an intented behavior and needs to be
solved