Implement LoginUserComponent as SubComponent
    As it is mentioned in the previous commit, the state remains after
logout.To solve the unintended behavior, anot component should be created.
    The created component should be specific to logged in user and also
should access the instances of ApplicationComponent.
    So, the SubComponent comes to rescue. When defining a SubComponent
it should be annotated with scope. In this example it would be ActivityScope.
    The injector functions is required to move into the LoginUserComponent.
Keeping them in the same Component will allow us to share the instances over
them. Additionally, the LoginUserComponent is marked with @SubComponent annotation
and related module is created and passed to the ApplicationComponent to
make ApplicationComponent aware of.
    Furthermore, HomeViewModel is marked with @ActivityScope annotation.
As you realized, it is the same scope with the SubComponent. It is working
as singleton but only in the LoginUserComponent.
    Lastly, When the HomeActivity is finished the instances attached to
LoginUserComponent is also deleted. Thus, the state will be reset after
logout action.