Provide dependency via provider interface
    When you have a class that is not created by dagger like AppInitializers
and it needs a dependency from the grap, the dependency could be provided
via an interface that has public getter functions for requested instances.

    In this example TimberInitializer depends TimberReleaseTree and getting
it over ApplicationComponent thanks to provider function in TimberInitializerDependencies.

    In addition, there is no requirement for public function to extract
an interface, what it means is, it can be directly put in the ApplicationComponent.
However, it will make the Component a god class and hard to follow. Creating
new interfaces for different classes can be thought as best practice.